requirements.txt
python-telegram-bot
requests
sqlite3
pyjq